Support nesting request metrics under a sub dictionary
In some cases you don't want the fields of the request metrics directly
in the root of the dictionary.

For example, when sending general application logs *and* request metric
logs to the same ES index it's good to be able to to nest the request
metrics as a separate object.

This commit adds 1 extra config option `NestKey` holding the name of
a key, that when set will nest the request metric fields in a sub
dictionary under the key name.